+++ unicorn_binance_websocket_api/unicorn_binance_websocket_api_connection.py
@@ -79,7 +79,8 @@ class BinanceWebSocketApiConnection(object):
                 except KeyError:
                     pass
                 self.handler_binance_websocket_api_manager.stream_is_crashing(self.stream_id, str(uri['msg']))
-
+                if self.handler_binance_websocket_api_manager.throw_exception_if_unrepairable:
+                    raise ValueError("stream_id " + str(self.stream_id) + ": " + str(uri['msg']))
                 sys.exit(1)
         except KeyError:
             pass
@@ -194,20 +195,14 @@ class BinanceWebSocketApiConnection(object):
         except websockets.exceptions.ConnectionClosed as error_msg:
             logging.critical("binance_websocket_api_connection->__aexit__(*args, **kwargs): "
                              "ConnectionClosed - " + str(error_msg))
-            if self.handler_binance_websocket_api_manager.is_stop_as_crash_request(self.stream_id) is True:
-                self.handler_binance_websocket_api_manager.stream_is_stopping_as_crash(self.stream_id)
-                sys.exit(1)
-            else:
-                self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
+            self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
             if self.handler_binance_websocket_api_manager.is_stop_request(self.stream_id) is False and \
                     self.handler_binance_websocket_api_manager.is_stop_as_crash_request is False:
                 self.handler_binance_websocket_api_manager.set_restart_request(self.stream_id)
             sys.exit(1)
 
     async def close(self):
-        if self.handler_binance_websocket_api_manager.is_stop_as_crash_request(self.stream_id) is True:
-            self.handler_binance_websocket_api_manager.stream_is_stopping_as_crash(self.stream_id)
-        else:
+        if self.handler_binance_websocket_api_manager.is_stop_as_crash_request(self.stream_id) is False:
             self.handler_binance_websocket_api_manager.stream_is_stopping(self.stream_id)
         logging.info("binance_websocket_api_connection->close(" + str(self.stream_id) + ")")
         try: