+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Semantic Versioning](http://semver.org/).
 
 ## 1.11.0.dev (development stage/unreleased/unstable)
+### Changed
+- loglevel connection().send() loglevels from error to critical.
 ### Fixed
 - `OSError` exception for `self.monitoring_api_server.start()` if its already started